# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Jiffoo Mall API",
    "description": "E-commerce System",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      }
    },
...
  },
  "paths": {
...
    "/api/admin/plugins/marketplace": {
      "get": {
        "summary": "Get marketplace plugins",
        "tags": [
          "admin-plugins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/plugins/marketplace/search": {
      "get": {
        "summary": "Search marketplace plugins",
        "tags": [
          "admin-plugins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/plugins/marketplace/{slug}": {
      "get": {
        "summary": "Get marketplace plugin details",
        "tags": [
          "admin-plugins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/plugins/categories": {
      "get": {
        "summary": "Get plugin categories",
        "tags": [
          "admin-plugins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/plugins/installed": {
      "get": {
        "summary": "Get installed plugins list",
        "tags": [
          "admin-plugins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/plugins/": {
      "get": {
        "summary": "Get installed plugins list",
        "tags": [
          "admin-plugins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/plugins/{slug}/install": {
      "post": {
        "summary": "Install a plugin",
        "tags": [
          "admin-plugins"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/plugins/{slug}": {
      "get": {
        "summary": "Get plugin state",
        "tags": [
          "admin-plugins"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/plugins/{slug}/enable": {
      "post": {
        "summary": "Enable a plugin",
        "tags": [
          "admin-plugins"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/plugins/{slug}/disable": {
      "post": {
        "summary": "Disable a plugin",
        "tags": [
          "admin-plugins"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/plugins/{slug}/config": {
      "put": {
        "summary": "Update plugin config",
        "tags": [
          "admin-plugins"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/plugins/": {
      "get": {
        "summary": "Get Installed Plugin List",
        "tags": [
          "Plugins - Public"
        ],
        "description": "Get list of locally installed plugins (no authentication required)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "category",
            "required": false,
            "description": "Filter by category"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "enabled",
            "required": false,
            "description": "Filter by enabled status"
          },
          {
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/plugins/{slug}": {
      "get": {
        "summary": "Get Plugin Details",
        "tags": [
          "Plugins - Public"
        ],
        "description": "Get detailed information about a locally installed plugin (no authentication required)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true,
            "description": "Plugin slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "development server"
    }
  ]
}